Create audio and video elements in constructor
This helps with an issue in Chrome: If the page is not in the foreground
then the custom element is not connected to the DOM even though the js
is still executed that triggers ICE negotiation.

Like this, the audio and video elements are just attached in the
onConnected callback, but created already in the constructor.