Fixes for Chrome 83 when the page is in the background
If the webrtc-media element was not yet connected to the DOM, then
the ICE negotiation still happens and the audio/video tracks are
added invisibly. Once the page is in focus again, the streams will
be visible to the user.

Happens e.g. if the user is on another tab or using fullscreen mode
in another application.