Create the RTCPeerConnection inside webrtc-media `connectedCallback`
The `connectedCallback` now emits an event that transforms the user from
a `UserWithoutPeerConnection` to a `User`.
If the remote peer sends an SDP offer, an answer is only created after
this transformation.

This solves the problems in Chrome where `connectedCallback/1` is not
triggered when the page is in the background in a more elegant way that
does not rely on a mutex that is read in two different files.